% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWorld.R
\name{makeWorld}
\alias{makeWorld}
\title{Make world}
\usage{
makeWorld(size = 100, range = 5, nclasses = 2,
  habitatColors = c("#fff200", "#1db125"), returnDF = FALSE, main = NULL)
}
\arguments{
\item{size}{how big should each side of the the square world be in pixels?}

\item{range}{how spatially autocorrelated do you want the data?  Higher values are more coarsely autocorrelated.}

\item{nclasses}{how many classes (e.g. habitats) do you want to produce?}

\item{habitatColors}{a vector of colors for the plot}

\item{returnDF}{do you want to return the dataframe with the raw variables or not?}

\item{main}{a title for the plot}
}
\description{
Make world
}
